Refactor to use more recognisable  controller pattern in favor of handler (which in Go has connotations of middleware)
